refactor(host)!: remove ctl_nats from host_config, introduce policy trait
